Ground reviews + preloader + error boudary

+++ Backend/Admin Portal/src/controllers/cityController.js
@@ -32,8 +32,13 @@ const getAllCities = async (req, res) => {
             return res.status(404).json({message: "Country not found."}) // Not Found
         }
 
-        const cities = country.cities;
-        res.status(200).json(cities);
+        const page = parseInt(req.query.page) || 1;
+        const limit = parseInt(req.query.limit) || 10;
+
+        const skip = (page - 1) * limit;
+
+        const cities = country.cities.slice(skip, skip + limit);
+        res.status(200).json({page, totalCities: country.cities.length, totalPages: Math.ceil(country.cities.length/limit), cities});
     } catch (error) {
         res.status(500).json({message: 'Unable to get cities.'});
     }
@@ -47,9 +52,8 @@ const getCity = async (req, res) => {
         }
 
         const city = country.cities.find((city) => city.cityId === req.params.id);
-
         if (!city) {
-            return res.status(404).json({message: 'City not found or not active.'});
+            return res.status(404).json({message: 'City not found.'});
         }
 
         res.status(200).json(city);
@@ -60,16 +64,26 @@ const getCity = async (req, res) => {
 
 const updateCity = async (req, res) => {
     try {
+        const country = await Country.findOne({countryId: req.params.countryId}).populate('cities');
+        if (!country) {
+            return res.status(404).json({message: 'Country not found.'});
+        }
+
+        const city = country.cities.find((city) => city.cityId === req.params.id);
+        if (!city) {
+            return res.status(404).json({message: 'City not found.'});
+        }
+
         //disallowing updating cityId which are supposed to stay unique
         if(req.body.hasOwnProperty('cityId') && req.body.cityId!=req.params.id){
             return res.status(400).json({message: "Changing city id is not allowed."});
         }
 
-        const city = await City.findOneAndUpdate({cityId: req.params.id}, req.body, {runValidators: true});
+        city.cityName = req.body.cityName || city.cityName;
+        city.image = req.body.image || city.image;
+        city.status = req.body.status || city.status; 
 
-        if (!city) {
-            return res.status(404).json({ message: "This city does not exist."});
-        }
+        await city.save();
 
         const updatedCity=await City.findOne({cityId: req.params.id});
         res.status(200).json({message: "City successfully updated!", updatedCity});
@@ -80,17 +94,26 @@ const updateCity = async (req, res) => {
 
 const deleteCity = async (req, res) => {
     try {
-        const city = await City.findOneAndDelete({cityId: req.params.id});
-        
-        if (!city) {
-            return res.status(404).json({ message: "This city does not exist."});
+        const country=await Country.findOne({countryId: req.params.countryId}).populate('cities');
+        if(!country){
+              return res.status(404).json({message: "Country not found."})
         }
 
-        res.status(200).json({message: "City successfully deleted!", city});
+        const cityIndex = country.cities.indexOf(country.cities.find(city => city.cityId === req.params.id));
+        if (cityIndex === -1) {
+            return res.status(404).json({message: 'City not found in this country.'});
+        }
+
+        country.cities.splice(cityIndex, 1);
+        await country.save()
+
+        const city=await City.findOneAndDelete({cityId: req.params.id});
+
+        res.status(200).json({message: 'City deleted successfully', city});
     } catch (error) {
-        res.status(500).json({ message: error.message});
+        res.status(500).json({message: error.message});
     }
-}
+};
 
 module.exports={
     addCity,
